将协议层从拉格朗日迁移到了Camille, 是c#的mirai-http实现。

+++ Meow/Core/Model/Base/DatabaseRecordBase.cs
@@ -1,5 +1,4 @@
 ﻿using FreeSql.DataAnnotations;
-using Newtonsoft.Json;
 
 namespace Meow.Core.Model.Base;
 
@@ -12,7 +11,7 @@ public abstract class DatabaseRecordBase
     /// 数据库id
     /// </summary>
     [Column(IsIdentity = true, IsPrimary = true)]
-    public int DbId { get; set; }
+    public long DbId { get; set; }
 
     /// <summary>
     /// 是否被删除
