+++ Meow/Plugin/NeverStopTalkingPlugin/Models/ForbiddenWordRecord.cs
@@ -9,7 +9,7 @@ public class ForbiddenWordRecord : DatabaseRecordBase
 {
     public ForbiddenWordRecord() { }
 
-    public ForbiddenWordRecord(uint recordUserId, string forbiddenWord)
+    public ForbiddenWordRecord(long recordUserId, string forbiddenWord)
     {
         RecordUserId = recordUserId;
         ForbiddenWord = forbiddenWord;
@@ -18,7 +18,7 @@ public class ForbiddenWordRecord : DatabaseRecordBase
    /// <summary>
    /// 记录者id
    /// </summary>
-   public uint RecordUserId { get; set; }
+   public long RecordUserId { get; set; }
 
    /// <summary>
    /// 违禁词
