+++ Meow/Plugin/NeverStopTalkingPlugin/Models/NstTriggerRecord.cs
@@ -7,7 +7,7 @@
 /// <param name="Uin">群|用户 标识号</param>
 /// <param name="TriggerFailedCount">触发失败的次数, 达到触发条件 但是没有取到消息or其他需要触发但没触发的情况</param>
 /// <param name="LastTriggered">上次触发的时间</param>
-public record NstTriggerRecord(BagOfWordType BagOfWordType, uint Uin, int TriggerFailedCount, DateTime? LastTriggered)
+public record NstTriggerRecord(BagOfWordType BagOfWordType, long Uin, int TriggerFailedCount, DateTime? LastTriggered)
 {
     /// <summary>触发失败的次数, 达到触发条件 但是没有取到消息or其他需要触发但没触发的情况</summary>
     public int TriggerFailedCount { get; set; } = TriggerFailedCount;
